==> scripts/06_creacion_de_dashboard.rmd <==
---
title: "Análisis de la relación entre movilidad y muertes por COVID-19 en Argentina"
output: 
  flexdashboard::flex_dashboard:
    vertical_layout: scroll
    source_code: embed
    social: menu
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(warning = FALSE)
```

Row {.tabset .tabset-fade}
-------------------------------------

### Movilidad vs muertes acumuladas

```{r}
mov_vs_muertes_acum
```

### Movilidad vs muertes diarias

```{r}
mov_vs_muertes_diarias
```

### Muertes diarias agrupadas por mes

```{r}
boxplot_muertes_diarias_por_mes
```

### Tabla de datos

```{r}
tabla_dinamica_argentina
```
